# Conflicts:
#	RR-Inuring/src/main/java/com/scor/rr/service/InuringGroupingService.java

+++ RR-Backend/src/main/java/com/scor/rr/domain/dto/TargetBuild/AssignTagToPltsRequest.java
@@ -7,7 +7,7 @@ import java.util.Set;
 public class AssignTagToPltsRequest {
     public Set<Tag> selectedTags;
     public Set<Tag> unselectedTags;
-    public Set<Integer> plts;
+    public Set<Long> plts;
     public Long wsId;
     public Integer userId;
 }
