+++ RR-Backend/src/main/java/com/scor/rr/repository/TargetBuild/PLTDetailSummaryRepository.java
@@ -3,5 +3,5 @@ package com.scor.rr.repository.TargetBuild;
 import com.scor.rr.domain.TargetBuild.PLTDetails.PLTDetailSummary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PLTDetailSummaryRepository extends JpaRepository<PLTDetailSummary, Integer> {
+public interface PLTDetailSummaryRepository extends JpaRepository<PLTDetailSummary, Long> {
 }
